Move generation logic into its own module
